Add root node presolve using Papilo (#234)
This PR implements:
- Adapters to presolve/postolve before/after running the solver
- Fetching of Papilo headers through Cmake
- System install of boost for pip wheel
- Runtime boolean parameter to enable the feature (true by default for MIP, false for LP). 

Notes:
- LP is off by default due to some presolvers not supporting dual post solve. These presolvers can be disabled but reduction quality is impacted.
- Post-solve status can appear as fail due to absolute/relative tolerances. Papilo uses absolute tolerance during post solve checks. Advice is to set relative tol in cuOpt to a very small value.
- TBB is disabled. To be enabled in a follow up PR for parallel presolve.

Fixes #277 
 
Credits to: @rg20

Authors:
  - Hugo Linsenmaier (https://github.com/hlinsen)
  - Rajesh Gandham (https://github.com/rg20)
  - Ramakrishnap (https://github.com/rgsl888prabhu)
  - Trevor McKay (https://github.com/tmckayus)

Approvers:
  - Alice Boucher (https://github.com/aliceb-nv)
  - Ramakrishnap (https://github.com/rgsl888prabhu)
  - Robert Maynard (https://github.com/robertmaynard)

URL: https://github.com/NVIDIA/cuopt/pull/234

+++ cpp/src/linear_programming/cuopt_c.cpp
@@ -437,7 +437,7 @@ cuopt_int_t cuOptGetVariableTypes(cuOptOptimizationProblem problem, char* variab
              variable_types.size(),
              problem_and_stream_view->stream_view);
   problem_and_stream_view->stream_view.synchronize();
-  for (int j = 0; j < variable_types_host.size(); j++) {
+  for (size_t j = 0; j < variable_types_host.size(); j++) {
     variable_types_ptr[j] =
       variable_types_host[j] == var_t::INTEGER ? CUOPT_INTEGER : CUOPT_CONTINUOUS;
   }
